Module rust_dns_https::worker

Reexports

use std::prelude::v1::*;

use error::Error;

use structs::APIResponse;

use dns_parser::{Packet, Question, QueryType, Builder, Type, QueryClass, Class, ResponseCode};

use hyper::{Url, Client};

use hyper::header::{Connection, Host};

use serde_json;

use std::io::Read;

Constants

GOOGLE_IP
The IP address of dns.google.com

Functions

build_response
Builds a response given a packet, and returns the bytes.

make_request
Sends an API request to GOOGLE_IP and parses the result into an APIResponse to return.

remove_fqdn_dot
Workaround for dns_pasrser, this is done since dns-parser improperly formats fqdns by removing the trailing dot.

translate_question
Translates a DNS question into a Google API Request